Butterfly/butterflyApp/consumers/slack-consumer/requirements.txt
kafka-python==1.4.6
mongoengine==0.17.0
slackclient==1.3.1
slack-bot==0.0.7
websocket-client==0.47.0